requirements.txt
numpy>=2.0.0,<2.2.0
scipy>=1.11.4
pandas>=2.0.0
scikit-learn==1.6.1
joblib>=1.3.2
fastapi==0.95.1
uvicorn==0.21.1
kafka-python==2.0.2
pymongo==4.11.3
requests==2.31.0
xgboost>=2.0.0
prometheus-client
prometheus-fastapi-instrumentator

